Create a bare app to demo a problem
This is an application which I expect to work with various plugins for
Vim by Tim Pope, but does not. If the application works as expected for
others, then it is likely indicative of a problem with my configuration,
though I'm not entirely sure where to begin figuring out how I broke
everything.

It's my fervent hope, then, that I'm simply misunderstanding the
documentation in some way, and there's a quick and easy step I missed.